iQEcBAABCgAGBQJW21uAAAoJEGDZx/EBlwS0XYMIAIUQ0HjYeQlVvlnPvgAgg9NF  RIr72sXje8f+MjhVrWS4YPd7cxIso1TcKP7PCOSP0+s4/ZpaHG1LWEVOEhLjTRF/  YJ2lX0kcfJoAVozTgjkUPqJtNOvfixZT/Zq7mWyc2yDmLKcabJGuBcFJcQzyZWWv  e+UkhgoTUPBwjU7UIcA/4TorVtPmjUs9dshMOFfiRGFR4o2sYK6PFVcairKgRN0P  fqugg50NqIMppd/HT4+jtjEiS5bV3CWqwPEaAGDKTsEity6IGWFvrg6+DQ/6taNQ  2ZavU93AiZaKaE5QB3etPi6/sh2Qojy23F9M7U6ju0OsoRxatKHBOhnY4JGkN/A=  =1Oxy  -----END PGP SIGNATURE-----
Support Ruby 2.2.0 and 2.3.0

Judging from the test suite, the only issue in supporting Ruby 2.2.0+
was with the `mysql2` gem. Updating this to the latest 0.3.x allows
Lobsters to run with newer Rubies.

We can't update to the absolute latest `mysql2` gem without first
updating Rails, which is why I used the pessimistic version constraint
for `mysql2`.